Apprestend

+++ src/main/java/app/business/controllers/rest/ProductTypeRestController.java
@@ -36,7 +36,6 @@ public class ProductTypeRestController {
 	UserService userService;
 	@Autowired
 	OrganizationRepository organizationRepository;
-	
 	@Autowired
 	ProductService productService;
 	
@@ -57,7 +56,7 @@ public class ProductTypeRestController {
 				{
 					map.put("id", Integer.toString(product.getProductId()));
 					map.put("name", product.getName());
-					map.put("quantity", Integer.toString(product.getQuantity()));
+					map.put("quantity", Float.toString(product.getQuantity()));
 					map.put("unitRate", Float.toString(product.getUnitRate()));
 					map.put("imageUrl", product.getImageUrl());
 					Listmap.add(map);
